feat: 系统总/可用内存读取 适配docker容器，硬件信息页面增加GC数据显示

+++ src/Admin/ThingsGateway.Admin.Application/Job/Hardware/HistoryHardwareInfo.cs
@@ -21,7 +21,7 @@ public class HistoryHardwareInfo
 
     /// <inheritdoc/>
     [SugarColumn(ColumnDescription = "内存")]
-    public int MemoryUsage { get; set; }
+    public UInt64 MemoryUsage { get; set; }
 
     /// <inheritdoc/>
     [SugarColumn(ColumnDescription = "CPU使用率")]
